requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain_community
langchain_core
langchain_ollama
chromadb
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
python-dotenv==1.0.0